List item no longer links
The list item for the vm no longer links to the vm page but instead the left side of the split button links. The right side of the split button is the dropdown for the vm controls. The left side when clicked changes the URL but the browser stays on the same tab instead of switching to the vm tab. Working on this issue...